+++ test/UnixFactAttribute.cs
@@ -1,0 +1,14 @@
+namespace RunScript;
+
+using System.Runtime.InteropServices;
+
+public class UnixFactAttribute : FactAttribute
+{
+    public UnixFactAttribute()
+    {
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+        {
+            Skip = "Ignored on Windows";
+        }
+    }
+}