+++ test/UnixTheoryAttribute.cs
@@ -1,0 +1,14 @@
+namespace RunScript;
+
+using System.Runtime.InteropServices;
+
+public class UnixTheoryAttribute : TheoryAttribute
+{
+    public UnixTheoryAttribute()
+    {
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+        {
+            Skip = "Ignored on Windows";
+        }
+    }
+}